Split query/strategy.rs tests module into its own file
